flask: let dom0 be a dm for PVH uivm.
The surfman/drm-plugin graphic stack needs to change the caching
attributes of the framebuffer region. This is done through
xc_domain_pin_memory_cacheattr which requires the dm permission.

This is only applied to uivm_t for now as xenfb2 is not used or
conveniently usable in a regular PVH guest.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>